Remove PaintController::SetPossibleBackgroundColor()
It was to let the platform painting code know the background color of
a table row or a table section, to let the containing composited layer
has a proper background color (used as placeholder when we don't have
enough time to rasterize a tile scrolled into view). However,
LayoutNGTables doesn't call it and we haven't received any bug report,
so we are removing the function to simplify code. The backgroun color
containing a table will be calculated merely based on painted results.

Bug: 1242202
Change-Id: I745ac55da029dc7a4474762ce978b316d1c9c19e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3257537
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938140}
NOKEYCHECK=True
GitOrigin-RevId: e63918abd9dc73f1b8a641e51f1816a556636851