Fix SignedCertificateTimestamp.timestamp documentation
The value is reported and consumed as a millisecond count rather than a
second count, so it is not a TimeSinceEpoch. Ideally it would be a
TimeSinceEpoch, but this would require a protocol change. For now, at
least make the documentation match reality.

Change-Id: I4c818903338aad5971bb882894ac90f278f31c67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3260453
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938437}
NOKEYCHECK=True
GitOrigin-RevId: b66427805865992bd90a234fca496c45534c337c