Use net::SSLInfo and friends in Blink
This CL reduces the amount of boilerplate needed to route TLS features
from the net stack to DevTools. There is something to be said for not
leaking this to the renderer in the first place, but that would require
a DevTools protocol change. This CL does not address that, but fixing it
could further reduce boilerploate. (In some configurations, the server
certificate we observed is PII. It's not web-platform-exposed, but it
seems odd to send non-platform-exposed things to the rendewrer.)

Blink currently has SecurityDetails/WebSecurityDetails, which contain a
slightly different version of the information in net::SSLInfo. These are
only used by DevTools. Due to historical reasons (before the repos were
merged, it was difficult to call into //net's logic to interpret TLS
codepoints), this information is stringified fairly early. This meant:

1. DevTools-specific display logic was split between ResourceResponse
   and InspectorNetworkAgent.

2. Blink retained string versions of codepoints which are much more
   compactly described with integers and enums in //net.

Meanwhile, the navigation stack was refactored, and now SecurityDetails
for navigation requests must be routed completely differently, up in
//content/browser. There, network_handler.cc has logic to directly
convert from net::SSLInfo to the DevTools protocol message.

Align the two paths and have Blink store a net::SSLInfo, instead of
maintaining parallel WebSecurityDetails and SecurityDetails structures.
This avoids needing to add each field two extra structs and plumb
between the two. It also stores a more compact representation.  Note
that net::X509Certificate is backed by CRYPTO_BUFFER, which is already a
certificate-specific AtomicString-like BoringSSL type. So retaining a
net::X509Certificate means we already get the memory savings of
AtomicString.

Unfortunately, we cannot quite deduplicate the Blink and content copies
of this logic, because the DevTools bindings are different between the
two. However, now the logic is at least analogous.

Since, as a result, we're losing some tests of the Blink half of the
logic (web_url_loader_unittest.cc checking different SAN combinations is
no longer relevant at that layer), I've added some new tests in
devtools_protocol_browsertest.cc, next to the very similar (but slightly
different) CertificateSecurityState test. These tests cannot use the
usual Blink machinery because they require configuring specific TLS
servers, so I've moved them up the stack.

Bug: none
Change-Id: I0b66f6c7e56395e1df2579103fe0540f9798d3ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3258063
Commit-Queue: David Benjamin <davidben@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938477}
NOKEYCHECK=True
GitOrigin-RevId: 316373d9a7d14dd1c4c70aa455dbeb3e3cdeb49b