Unwind SecurityStyleExplanations
DevTools no longer uses this. This causes the securityStateChanged event
(which DevTools now ignores) to always have an empty explanations field
and omitted summary field.  For now, this leaves
WebContentsDelegate::GetSecurityStyle in place, but it's only used by
this same securityStateChanged. As follow-up work, we should unwind a
bit more of this.

This uncovered a dependency issue in //headless (depending on the
//components/policy generation process and forgetting to link in the
generated library itself), so I've fixed that here too.

Tested manually that the DevTools security panel still works on various
badssl.com pages.

Bug: 1262378
Change-Id: I955c6f0fc893bbafe92f50a5983389e2f7a6a45a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3247295
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938488}
NOKEYCHECK=True
GitOrigin-RevId: e9084f62ea22b0ca06db5deaff87ff92bcbce58d