[FlexNG] Refactor in GiveItemsFinalPositionAndSize()
Adjust GiveItemsFinalPositionAndSize() so that we do a full loop of
all flex items in the first parent fragment to allow certain
item information to get propagated that eventually won't be available
in the next layout pass. This propagation was refactored into a new
method called PropagateFlexItemInfo().

This is in preparation for a larger refactor meant to optimize flex
fragmentation and should not result in any functional changes.

Bug: 660611
Change-Id: I99b941e3af52e713e05883ff2cae7ebcd49fe478
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3248274
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938541}
NOKEYCHECK=True
GitOrigin-RevId: 1da45d89dfaa3346c2a494990bae17debbd8f9ab