Fetch code cache earlier in NavigationBodyLoader
The results from the NavigationThreadingOptimizations experiment have
shown that decreasing the time we wait for the code cache can have a
large affect on FCP/LCP. This change attempts to fetch the code cache
much earlier in commit so it will hopefully be ready when needed.

Note if the main contributor to the code cache response time is the
queue time of the response task in the renderer, this probably won't
help much. If experiment results don't show much of a speedup, that's
the next place to optimize.

Bug: 1266146
Change-Id: Ie8d702b467ce7455a65ade1ca70d30b2def22870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3258419
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938553}
NOKEYCHECK=True
GitOrigin-RevId: 038cd96142d384c0d2238973f1cb277725a62eba