Refactor partitioned service worker 3p window for reusability
Future partitioned service worker tests are likely to follow the
pattern of opening a window in a third-party context and then
having that window spawn an iframe that is same origin to the test.

This change modifies the existing 3p window file to allow the test
to specify which url the window should iframe. Allowing for reuse
between tests.

Bug: 1246549
Change-Id: I07726baad26fc6ac54538344c58b59dfcfe7f6a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3259316
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938562}
NOKEYCHECK=True
GitOrigin-RevId: bb4c119b622fa889ef6bab395e29fb99d2ff4a94