DevTools: Emulate `forced-colors` media feature
`forced-colors` CSS media feature allows developers to detect if the UA
has enabled a limited color palette and forces the page to conform.

This patch provides backend support in order to add UI to the Rendering
drawer and a Command Menu option that allows a developer to toggle a
preset for `forced-colors: active` (see crrev.com/c/2324875).

It does so by creating a pre-canned set of system colors to be forced
when emulating forced colors mode, then restores the original on
completion.

Bug: chromium:1130859
Change-Id: I9931e5e0e4c85f01b138b1c70c660b43bf114a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333043
Commit-Queue: Kalon Hinds <kahinds@microsoft.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938623}
NOKEYCHECK=True
GitOrigin-RevId: c29c95b10310654f06451375eb327a1313fa504e