Add option to proto_library to *always* use source_sets.
The proto_library GN template will usually generate
static_libraries for the generated C++ code, but it supports
an option to force the library to be a source_set in
the component build; this can be useful if you need to export
the symbols from a shared library containing the generated code.

However, this option only works in the component build. If you're
actually building a shared_library/loadable_module as the linkable
object in a release build, this won't do you any good.

This CL adds an option to force the template to use a source_set
unconditionally (i.e., regardless of whether this is a shared_build
or a static_build).

Bug: 1266935
Change-Id: I9e22414b79cb231994f8b8df9491dd2b9ce0e49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3259427
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#938652}
NOKEYCHECK=True
GitOrigin-RevId: 0b646dbffbca733e3fa24128bab1472c752a032f