Enable OneCopyCanvasCapture on all platforms
This will only take effect on platforms that also have the
TwoCopyCanvasCapture feature enabled.

The old path copies from a WebGL DrawingBuffer to a StaticBitmapImage
to a VideoFrame.

The new path copies from the WebGL DrawingBuffer to a VideoFrame.

Future work can extend this to WebGPU and other types of canvas.

Bug: 1207114
Change-Id: Id6b5ed93d982370934f5e0c5848a4549bef47ad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3258053
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938672}
NOKEYCHECK=True
GitOrigin-RevId: aca9bf707a0b9662330142aeaf53619cc5a4cffb