[mq-4] Special handling for empty input case
We currently have tests that expect an empty input to parse
as an empty MediaQuerySet. This CL adds a special branch for this
at the start of MediaQueryParser::ParseImpl. Dealing with this case
early makes further simplifications easier to understand.

I believe we already effectively have a special case for this
through the else-if branch removed in this CL, except that you
need to stare at the state machine for a long time to see it for
what it is. I will attempt to show that if we handle the empty case
early, the else-if branch will always be false, and therefore empty-
input handling was its only purpose:

The else-if branch can only be true if we're at range.AtEnd().
The only way for the ReadRestrictor function to observe that range, is
through the special and final empty-range call to ProcessToken in
ParseImpl.

There are two ways to be in state=kReadRestrictor at the time
of the final ProcessToken call:

 1) Not changing the initial state, i.e. empty input. However, as of
    this CL we catch that case early, and never reach the final
    ProcessToken call.
 2) By state=kReadRestrictor being set during the state machine
    execution. This happens in two places:
    - FinishQueryDataAndSetState
    - SkipUntilComma
    In both cases, if state=kReadRestrictor is set, then we *also*
    call AddMediaQuery. This means the !query_set_->QueryVector().size()
    part of the else-if can never be true.

On a related note, it is no longer possible to get kReadMediaNot as an
end state, since that state is only set initially and then ever again.

Bug: 962417
Change-Id: Ifdee0e8e634a78ceaad58f31b6cb6fb76ee2df5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3250744
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#938808}
NOKEYCHECK=True
GitOrigin-RevId: d996f664ad51d98044704e93f574f4d3488a637a