Remove LargestContentfulPaint1
This CL removes the ukm reporting as well as the variables within the
paint timing detectors that keep track of the old LCP version. The
mojom attributes are kept because there are several incoming LCP changes
so it's likely that it could be useful to keep the plumbing for those
future experiments. In a followup, the paint timing detectors can be
simplified as they now only need to compute the new LCP version.

Bug: 1045640, 1265557
Change-Id: I1ab9f64ec89d1057e6cca15465c03ec55d24ce28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3256966
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938811}
NOKEYCHECK=True
GitOrigin-RevId: f4a8312700fcb62bec6070864a0d5b13091b609c