Add androidx_biometrics library.
License Review is not needed as this is the same source and license
of already approved code in same party. Size increase is not an issue
in this CL as it just allows the library to be built against in another
CL where size can be addressed.

Androidx biometrics would allow support in android below version R.
https://source.chromium.org/chromium/chromium/src/+/main:components/device_reauth/biometric_authenticator.h;l=40?q=biometric_authenticator.h

Bug: 1056372
Change-Id: I9b3bac97229129b6176f899289a8a486c2306339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3248123
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Ben joyce <bjoyce@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938860}
NOKEYCHECK=True
GitOrigin-RevId: 72740fdde70c8d72431bffcdfd74003b11f43a1e