[mq-4] Treat a final state of kReadRestrictor as an error
We currently explicitly call ProcessToken() on an empty range to
make it possible for handler functions to check for range.AtEnd()
and take some action based on that. This call has one of three
possible effects:

 1) kDone -> kDone (i.e. no effect).
 2) kReadRestrictor -> kSkipUntilComma
 3) kSkipUntilComma -> kReadRestrictor + AddMediaQuery

In case (2), this means that we add "not all" at the call-site after
ProcessToken(empty) has finished. This happens when there's a valid
query followed by a single final comma, e.g. "(color),".

In case (3), we add "not all" *during* SkipUntilComma, and then
prevent an additional erroneous "not all" from being added
by checking CurrentMediaQueryChanged(). This happens when there's
an invalid query followed by a single final comma, e.g. "invalid,".

If we instead remove ProcessToken(empty), then for case (2)
we can just handle kReadRestrictor identically to kSkipUntilComma,
and for case (3) the fact that we remain kSkipUntilComma will produce
the "not all" we previously produced inside the handler.

Bug: 962417
Change-Id: I75e1b41560a235e4c03877c0fec83a5f7a19810c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3253389
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#938879}
NOKEYCHECK=True
GitOrigin-RevId: 34015bacdb0845a226035749e1077c499419970f