EditContext: introduce character bounds
This CL implements the following APIs for character bounds which let
the author specify the bounding box of characters that are requested
by the browser. The info is used by the IME to display UX (ex. candidate
window for Japanese) at the correct position when in active composition.

EditContext spec:
https://w3c.github.io/editing/docs/EditContext/index.html

1. CharacterBoundsUpdateEvent [1], which informs the author a range
in which character bounds are needed by Text Input Service.

2. updateCharacterBounds() [2], which lets the author provide the
requested character bounds which will be cached in EditContext.

3. characterBoundsRangeStart, characterBounds() [3]
getter for cached values.

Previously if EditContext is active, we have a special logic in
TSFTextStore::GetTextExt to always return the selection
bounds (provided by EditContext) regardless of the requested range from
TSF (which is incorrect). This CL removes the special logic, modify
RenderWidgetHostViewAura::GetCaretBounds(), and implement
EditContext::GetCompositionCharacterBounds(), so that
when EditContext is active, the caret bound and the character bounds
can be correctly retrieved.

test: edit-context.html

[1]: https://w3c.github.io/editing/docs/EditContext/index.html#characterboundsupdateevent
[2]: https://w3c.github.io/editing/docs/EditContext/index.html#dom-editcontext-updatecharacterbounds
[3]: https://w3c.github.io/editing/docs/EditContext/index.html#dom-editcontext-characterbounds

Bug: 999184
Change-Id: I3a8a52b052afdaeeccbae57b877982c8885d1235
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3260767
Commit-Queue: Alex Keng <shihken@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938893}
NOKEYCHECK=True
GitOrigin-RevId: d50b5b8a78d3d4878c98b07af006a00ed65fddd0