[Blink] Add promise support to WebLocalFrame::RequestExecuteScript()
Add optional promise support to WebLocalFrame::RequestExecuteScript().
When specified, if the executed script evaluates to a single argument
of a v8::Promise, then the script executor will wait for the promise
to resolve and pass the result of the promise to the
WebScriptExecutionCallback. This avoids callers having to handle the
Promise-chaining themselves.

If a promise rejects, an empty v8::Local<v8::Value> is passed to the
callback (matching how we treat errors in scripts).

Use the new promise behavior in CommerceHintAgent, which previously
manually chain a method onto the returned promise.

Bug: 1195807
Change-Id: I07d7e8450d57ccb94cce1cfb1595cc490d90d166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3230010
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938948}
NOKEYCHECK=True
GitOrigin-RevId: 40d8e082e3f4f4f2322b9747e64ba434988e56da