[LCP] UKM indicating animated largest candidate.
This CL adds a UKM indicating if the largest candidate was an animated
image. This is related to [1], and will help us evaluate the metric
change when the flag is enabled. At the same time, this metric does not
depend on a flag.

UKM collection review [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3250690
[2] https://docs.google.com/document/d/1CjgK1vOOJ6erN1VBNb80QO9wmeEdqhmRsij_YJ7frXc/edit#

Bug: 1260953
Change-Id: Ia90825e33aa6200b68671b16af9ab691ae1c9f30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3245540
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939465}
NOKEYCHECK=True
GitOrigin-RevId: f7d296d7ac4d101314aa8cb0ed892080cb3ea3b2