Fenced Frames: RemoteFrame should also do replace-navigation
The initial CL for Fenced Frames' replace-only navigation changes
landed in
https://chromium-review.googlesource.com/c/chromium/src/+/3260313
but one of the flows not covered there was a cross-site frame's
navigation in the shadowDOM approach. This is because there are two
versions of NavigationShouldReplaceCurrentHistoryEntry. The last CL
updated the one in LocalFrame but not in RemoteFrame which this CL does.

Bug: 1267637
Change-Id: Iae70f00546a4bec06be60fb00185ef22011287b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268425
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939672}
NOKEYCHECK=True
GitOrigin-RevId: 08b7bde6863be64f6d1fe9853618b987652e953b