Let getElementById receive MaybeAtomicStrings for fast response.
Implements a new MaybeAtomicString class derived from AtomicString with
the same functionality as its parent, but which holds a null string if
the queried string wasn't already on the AtomicStringTable. With this,
getElementById (and potentially other v8 callbacks) can give a fast
response if the queried name is not in the string table without creating
a new AtomicString, thus avoiding an instance of refcount trashing.

Bug: 1083392
Change-Id: Ie256e11abbc430bebdf2844d22701bb85e8b5464
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3197846
Commit-Queue: Luis Fernando Pardo Sixtos <lpardosixtos@microsoft.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939683}
NOKEYCHECK=True
GitOrigin-RevId: 6acea4080a9019dd4d5e850a2b0a95efd9475db8