Add new method to release StringBuilder::string_ object.
Most instances of StringBuilder are used to construct a single String
object and disappear after StringBuilder::ToString() is called to get
it. This causes refcount trashing when Release is called on the
builder's string_ and AddRef is called on the returned one. This
behavior causes around 132 unnecessary refcount changes in the
CSSPropertyGetterSetter benchmark, and potentially more across blink.

This CL adds a StringBuilder::ReleaseString() method with the same
functionality as StringBuilder::ToString() but that releases the
string_ object to the caller. The CL by itself will not reduce refcount
trashing, but subsequent changes transitioning ToString() calls to
ReleaseString() will.

While in the area, I wrapped some of the internal logic of ToString()
and ToAtomicString() into a template to avoid repeated code.

Bug: 1083392
Change-Id: I4f05250be5ec6f2879ece0e3688c782e256e87ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3262371
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Luis Fernando Pardo Sixtos <lpardosixtos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#939701}
NOKEYCHECK=True
GitOrigin-RevId: d4f604f406aaac313cc29e2b4bc1d25f94256265