Add getAXNodeAndAncestors CDP command
This command takes a DOM node reference and retrieves the corresponding
AXNode and its ancestors up to the root. We use this command in the
The DevTools frontend lazily builds the accessibility tree. We use this
command to map DOM nodes to AXNodes that lives in parts of the tree that
has not been previously requested. Using the ancestor chain up to the
root we can then stitch the returned subtree together with the already
known part.

Bug: 1226486
Change-Id: I374765908358a88aaf2d21bc17acf814207dfc9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3253351
Commit-Queue: Johan Bay <jobay@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939724}
NOKEYCHECK=True
GitOrigin-RevId: cf39dcfaa19cd371596e9ce117be6e0ca0c86d07