[Deprecation Issue for Legacy Client Hints] (1) Add Feature {Redux}
I forgot that since we need to check for deprecation in the renderer and
browser we need to put shared code in services. This migrates that code
and adds a test for a core component (checking which hints are
deprecated and suggesting the alternate).

This CL is part of a series, search for the bracketed prefix to find them:
(1) Add Feature
(2) Implement (Chromium + HTML)
(3) Implement (Chromium + HTTP)
(4) Implement (DevTools)
(5) Enable by default
(6) Bake in

Bug: 1227043
Change-Id: I58d82785df4905b3cf5eb0a14ecb698eeefc4993
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3267826
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939932}
NOKEYCHECK=True
GitOrigin-RevId: 75182b830e1e4e5ed2f94870272f00fa92f742fc