Pre-paint: Only walk unfragmented OOFs once.
If the parent of an OOF is participating in a fragmentation context
which the OOF child isn't part of, we only need to walk it once. No
point in doing it for every inner fragmentainer (since the OOF isn't
contained within that fragmentation context).

Remove virtual/layout_ng_block_frag/ specific expectation for
fast/multicol/mixed-opacity-test.html . With this CL it will fail
identically to legacy - that's crbug.com/807395. Previously, with NG
block fragmentation, we'd paint correctly on top of the second column,
but not the first. Now we'll paint incorrectly in both, which is more
consistent, if nothing else. See tc3.html in crbug.com/1253797 (it will
now render identically in legacy and NG).

Bug: 807395, 1253797, 829028
Change-Id: I5a3d6862382df7da75093c3b9adaf1948cb50c8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3270330
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940058}
NOKEYCHECK=True
GitOrigin-RevId: ecdc6031361ceccf0b1623cc416cfcd16862a4e3