Window Placement: have ScreenDetailed support same quirks as Screen
This is a drive-by cleanup.  It seemed odd that Screen::width()
could return a different value from ScreenDetailed::width() if
the physical pixel quirk was enabled.

A followup to this patch is to move display_id_ into the Screen base
class so that ScreenDetailed doesn't need to override it.  However,
this patch needs to happen first to converge the behavior.

Existing browser tests cover regressions of the non-quirk use case
in ScreenDetailed and wpt tests cover the Screen versions of the quirk
use cases.  However, I don't think it would be worth trying to write
a browser test that exercises this particular quirks setting.
Testing for this code with this quirk can wait until after issue
1252062 is fixed (see notes there) and we can write a WPT (internal?)
test.

Also, while we're here, switch to using the nicer safe conversion
code from base rather than static casting.

Bug: 1267398
Change-Id: Idf447cc406cb24dc6b49b0220352860fb0541ea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3265187
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940108}
NOKEYCHECK=True
GitOrigin-RevId: 4c6f4c80dab39bfd8a08e4dec5bb591a41846fa1