RTCDataChannel: Ignore send failures
There is a possible race between send() and close events between the
WebRTC and Chrome layers.
In this case, the Chrome layer will try to send() over a data channel
with a pending remote close event. It is safe to ignore this error.
Queued data on a data channel is discarded on remote close(), so the
observable behaviour is correct.

Bug: 1260845, 1267579
Change-Id: I0194102d91293979caf3092d6deb61a00c82d6e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268746
Commit-Queue: Florent Castelli <orphis@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940132}
NOKEYCHECK=True
GitOrigin-RevId: df984d1a46a2c120df620e15c822ef6af9348df1