Skip DisplayLock forced updates when outside flat tree
It would seem that the style code can get in a bad state when we try to
set style bits in a node that's outside the flat tree during a forced
update. This patch skips setting those bits when the node to unlock is
outside the flat tree.

Fixed: 1266891
Change-Id: Idd6af0ad60f7151bad9c59c4996d539691ae7408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268447
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940136}
NOKEYCHECK=True
GitOrigin-RevId: efe767952d0105b608047601737e70e08208837f