Vector: Clear unused vector on move
Explicitly clearing allows the backing to be freed immediately. In the
non-garbage-collected case this is often just slightly moving it
earlier as the old backing would otherwise be freed in the destructor.
For the garbage-collected case this allows for freeing the backing
right away without introducing GC pressure.

This is an alternative way to fix the attached bug.

Bug: chromium:1260856
Change-Id: Ibc49449dbce395ae949e27fef9a502c120e38e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3270340
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940260}
NOKEYCHECK=True
GitOrigin-RevId: 974f0365ab488c6cab3cd8cbee165238873853ff