Support style invalidation of ':has()' in rightmost compound
This is the initial CL of supporting style invalidation with ':has()'.

To minimize complexity, we will focus on the following limited usage.
- Supports ':has()' in a rightmost, top-level compound
- Supports ':has()' argument starts with child or descendant combinator
- Only allow attribute or elemental selectors in argument

So the following will be ignored for now.
- Ignore ':has()' in a non-rightmost compound
- Ignore ':has()' in a logical combinations
- Ignore ':has()' argument starts with '~' or '+'
- Ignore all pseudos in ':has()' argument

To support this case, for a change on an element, we need to invalidate
ancestors.

Invalidate every ancestors for every mutations on every element is very
bad for performance.

So we need a way to skip these.
- irrelevant element
- irrelevant ancestor
- irrelevant mutation

To skip the 'irrelevant element' and 'irrelevant ancestor', this CL
introduces two flags in the ComputedStyle.
- AncestorsAffectedByHas   : Indicate that the element has an ancestor
                             affected by ':has()' state change
- AffectedByHas  : Indicate that the element can be affected by ':has()'
                   state change

During style recalculation, 'AffectedByHas' and 'AncestorAffectedByHas'
flag will be set when SelectorChecker tries to match any ':has()'
selector on an element. The 'AncestorAffectedByHas' flag is inherited
by children of the element.

<style> .a:has(.b) { color: red }</style>
<div>
  <div class="a">    <!-- AffectedByHas, AncestorAffectedByHas  -->
    <div>            <!-- AncestorAffectedByHas  -->
      <div>          <!-- AncestorAffectedByHas  -->
      </div>
    </div>
  </div>
  <div>
    <div>
    </div>
  </div>
</div>

Only for a mutation on the element with 'AncestorAffectedByHas', style
engine will walk up ancestors. And it invalidates style only when an
ancestor is 'AffectedByHas'.

Still we have invalidation for irrelevant mutation and invalidation on
irrelevant ancestors. We will handle those by separate CLs.

Bug: 669058
Change-Id: Ia7e607e01f9a070c0d69d89679b99efa0a104797
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3260677
Commit-Queue: Byungwoo Lee <blee@igalia.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940274}
NOKEYCHECK=True
GitOrigin-RevId: 4c1e19544a4900e00be4d3863bd644dfe32d1502