Simplify TextPaintTimingDetector
In this CL, we do the following changes:
* Only use TextRecords when we need to measure the paint time or for the
largest text or largest ignored text. For any other nodes, we only need
to know whether they have been recorded as candidates or not.
* Get rid of TextRecordsManager as we only keep track of a few set of
TextRecords.
* Change the semantics of the LCP information store to track the LCP2
candidate. This will help simplify the logic in PaintTimingDetector,
as before the record would represent the LCP1 candidate.
* Change the logic to ensure that text nodes get accurate paint times.
We keep track of a frame index and use that in order to determine which
of the queued nodes need to receive a paint time in the given callback.
This matches the implementation in the ImagePaintTimingDetector, so the
logics can be merged into the PaintTimingDetector in the future.

Bug: 1267368
Change-Id: Ia7ebac0fae27732bb269e2e77cda0de19eab8e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3265727
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940502}
NOKEYCHECK=True
GitOrigin-RevId: f9aa7679e65ab02057735e3243eef66d2a08d784