Fix DCHECK in child test procs when GTest assertions fail.
GoogleTest generally calls the UponLeavingGTest() method of an
OsStackTraceGetterInterface implementation when it runs code-under-test.
This is used when a stack trace is gotten to omit the stack frames that
precede the code under test, since they're just noise in the output.
This is all well and good, except when we use GoogleTest macros in child
processes where the code-under-test isn't run via a Run method in
GoogleTest. Rather than DCHECK in that case (which is not at all
helpful), let's simply include all stack frames in the output.

Bug: None
Change-Id: Ic8883526defaf1823c73e1a0e12f7003785dc32d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3275452
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940854}
NOKEYCHECK=True
GitOrigin-RevId: 0963fcf750c81fc9dd9e6fea5c4b6ff4c2076553