Do more cleanup in FontCache family name selection
Initial attempt was done in [1] but reverted in [2] because it was
breaking default serif rendering on Android. It turns out that it is
still necessary to handle FontDescription conversion for the family
name "-webkit-standard" on Android, so this CL restricts that
conversion to that case. This might be further simplified once [3]
is fixed.

This CL also adds back the DCHECK for "system-ui" which was
unintentionally removed in [2] as well as a non-regression test
for bug 1252383. It also removes the Linux-specific expectation added
in [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3114293
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3202092
[3] https://bugs.chromium.org/p/skia/issues/detail?id=12503

Bug: 1252383, 1241875
Change-Id: I64168a04123e22a101c4299cee88535c70e17bb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268597
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941116}
NOKEYCHECK=True
GitOrigin-RevId: 2e863a5bce60e604911e7cbd3b27ddc760cd6e5e