Changes to column-rule properties only require repaint.
Column rules don't take up space in layout, so it's not necessary to
relayout when such properties change.

This fixes paint/invalidation/multicol/column-rules-fixed-height.html
when LayoutNGBlockFragmentation is enabled. It used to fail because we
get a lot of cache misses when laying out a multicol container. The only
thing we're able to cache is unfragmented nodes, and even that is only
if there's no column balancing involved, in which case we'll miss
everything. See crbug.com/1269485

Bug: 829028
Change-Id: Iadc53ff7be86f99d8a2e2f9dc1e328bd6c2c5f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3275501
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941136}
NOKEYCHECK=True
GitOrigin-RevId: 5038dc9c34771d4953ce4737aa583136442ae25f