[GridFragmentation] Position items based on their row position.
Previously we'd decide if a grid-item should be placed in the current
fragmentainer based on their global position.

This was ok - but a better strategy is to try and keep them with their
row. This wasn't initially important, but as break avoidance is row
based, when we try to shift a row to a subsequent fragmentainer we'll
potentially end up with an item which has already been placed.

This has the side-effect that a grid-item with a negative block-start
margin, may be placed above the start fragmentainer edge - which seems
ok.

I added this as a tentative testcase.

Bug: 614667
Change-Id: I65486307fd72f8e5f8cb7f87ebce66c5569deedf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3276728
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941176}
NOKEYCHECK=True
GitOrigin-RevId: 6919d34cc2e0bbd0be82c7e9c3c167595268cf46