WebTransport: Use UseCounter::Count
Use UseCounter::Count in web_transport.cc instead of
ExecutionContext::CountUse. The difference that that the UseCounter
variant performs a null check and so can safely be used after the
ExecutionContext has been destroyed.

Add a test that it safe to access WebTransport getters even after
the frame in which the WebTransport was created has been destroyed.

BUG=1269525

Change-Id: I4af26f3e34adcfcf28bf39f99f6e694fb61d91c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3276626
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941183}
NOKEYCHECK=True
GitOrigin-RevId: 7560171a69a4432700e41e3c8d2f6d3650d95f5d