Introduce PageLifecycleStateUpdater to share code used in updating.
If our logic is correct, this should be a no-op. That would imply that
the state-data in browser/renderer gets out of sync.

The browser and renderer have difference classes to represent
PageLifecycleState:
- blink::mojom::PageLifecycleStatePtr
- mojom::blink::PageLifecycleStatePtr

so this has to be a template for it to work in both.

Bug: 1234634
Change-Id: If88398f594a108edac0dc4e8b701a79645b1d1fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3256289
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941575}
NOKEYCHECK=True
GitOrigin-RevId: d2cd6abf0be3bfe0b36c6fb8215b5ee94d17b588