Consider editing boundaries when determining caret bounds
In r938733 we landed a patch that added an extra argument in the
functions responsible of determining the care local and absolute
bounds. This new argument provides the behavior the caller wants to
deal with the DOM tree editing boundaries.

However, we deliberately ignored that argument in the functions to
compute the downstream and upstream equivalent positions, relying on
the default behavior of the MostForwardCaretPosition and
MostBackwardCaretPosition functions, which allow crossing the editing
boundaries.

Now with this change we consider the editing boundaries behavior passed
by the caller, which may lead to a general behavior change, depending
on how the caret related functions, mentioned before, are used.

This CL implements a few unit tests to verify how we are determining
the caret bounds obtained with different editing boundaries behavior.

Bug: 688015
Change-Id: I4fe4466b0512f42e1ba0990e3d53eb6929e7fb80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268124
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941670}
NOKEYCHECK=True
GitOrigin-RevId: 6a65d39ba49dc49d626d61388260dfa51831a035