Fix animations created with an invalid execution context.
Previously, the execution context was implicitly assumed to be valid in
the animation constructor. It is OK to run an animation without an
execution context. We simply can't invoke any JavaScript callbacks in
this case.  We already have safeguards if the execution context is
destroyed after the animation creation, but did not verify the validity
of the execution context when constructing the animation.

Bug: 1254444
Change-Id: I8f8022222988785f0b7f07a1b258989bc37117fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3278735
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941761}
NOKEYCHECK=True
GitOrigin-RevId: da480c329100f7af095322df8484c7af61aba321