[Region Capture] rewrite scrolling test
This patch updates the ProduceCropIdRegionCaptureData scrolling test to
no longer depend on calling the 'produceCropId' API, and renames the
test to ElementRegionCaptureData to reflect the change.

Instead of relying on the JavaScript API, now the test sets the crop id
directly on the Element in order to test the paint system changes.

Bug: 1269886
Change-Id: Ibe85f0c467eafd8bb391da9ebd30d3d774f22482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3282768
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941775}
NOKEYCHECK=True
GitOrigin-RevId: db87d8e8a0cefe994922e6c123051845f0fb8405